requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
Django==4.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
fiona==1.10.1
GDAL==3.8.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ijson==3.4.0
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.67
langchain-openai==0.3.27
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
mapbox-vector-tile==2.2.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openai==1.93.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pillow==11.3.0
protobuf==6.31.1
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.29
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyJWT==2.10.1
PyMuPDF==1.24.14
pyparsing==3.2.3
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
shortuuid==1.0.13
simplejson==3.20.1
simplekml==1.3.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
zstandard==0.23.0
